+++ src/components/pages/Order/Main/MainRightSide/AdminPanel/AddForm.jsx
@@ -1,14 +1,16 @@
-import { useContext, useState } from "react";
+import { useContext } from "react";
 import OrderContext from "../../../../../../context/OrderContext";
 import { getInputTextsConfig } from "./inputTextConfig";
 import { EMPTY_PRODUCT } from "../../../../../../enums/products";
 import Form from "./Form";
 import SubmitButton from "./SubmitButton";
+import { useSuccessMessage } from "../../../../../../hooks/useDisplaySuccesMessage";
 
 export default function AddForm() {
   // state
   const { handleAdd, newProduct, setNewProduct } = useContext(OrderContext);
-  const [isSubmitted, setIsSubmitted] = useState(false);
+  const { isSubmitted, displaySuccessMessage } = useSuccessMessage(2000);
+  // const [isSubmitted, setIsSubmitted] = useState(false);
 
   // comportements (gestionnaires d'évenements ou "event handlers")
   const handleSubmit = (event) => {
@@ -29,13 +31,6 @@ export default function AddForm() {
     setNewProduct({ ...newProduct, [name]: value });
   };
 
-  const displaySuccessMessage = () => {
-    setIsSubmitted(true);
-    setTimeout(() => {
-      setIsSubmitted(false);
-    }, 2000);
-  };
-
   const inputTexts = getInputTextsConfig(newProduct);
   // affichage
   return (
@@ -45,8 +40,8 @@ export default function AddForm() {
       onSubmit={handleSubmit}
       onChange={handleChange}
       isSubmitted={isSubmitted}
->
-        <SubmitButton isSubmitted={isSubmitted} />
+    >
+      <SubmitButton isSubmitted={isSubmitted} />
     </Form>
   );
 }
